rename image/image.go to image/vips.go
